moved the saveJob stuff to a method on Job instances
